+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+- 2021-12-08 RCSS-813 related:
+
+  - Add `ImagePainter::line()`.
+  - Add parameter `CustomInflationFunc func` to constructor of `CostmapInflator`.
+  - Add `CostmapInflator::setSeed()`.
+
 - 2021-12-06 Add Base64.
 - 2021-11-25 Add StringSlice::makeWithContentsOfFile().
 - 2021-11-25 Fix bug in StringSlice::trim().